+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   code using some predefined identifiers (`TEST`, `BENCH`, `IFDEF`)
   (#345, @NathanReb)
 + Add support for starred style comments (#347, @NathanReb)
++ Add explicit support for labeled tuples syntax (#349, @NathanReb)
 
 ## 1.9.0